// src/assets/Interfaces.tsx
export interface ILogoAsset {
    /**
     * Squared logo with black background
     */
    squared: string;

    /**
     * Serch TagLine in `#050404` color
     */
    tagBlack: string;

    /**
     * Serch TagLine in `#ffffff` color
     */
    tagWhite: string;

    /**
     * Serch Logo in `#050404` color
     */
    black: string;

    /**
     * Serch Logo in `#ffffff` color
     */
    white: string;
}

export interface IAppAsset {
    /**
     * Serch User/Guest App Design
     */
    user: string;

    /**
     * Serch Provider/Associate Provider App Design
     */
    provider: string;

    /**
     * Serch Business App Design
     */
    business: string;
}